refactor(AlertMailer): move to training subpackage
