+++ tests/unit_tests/utility/test_alert_mailer.py
@@ -1,7 +1,7 @@
 import pytest
 from unittest import mock
 
-from brisk.utility.alert_mailer import AlertMailer
+from brisk.training.alert_mailer import AlertMailer
 
 class TestAlertMailer:
     """Test class for AlertMailer."""
